+++ src/Plugins/Import/GpxImport/src/GpxParser.cpp
@@ -191,7 +191,7 @@ void GpxParser::parseWaypoint() noexcept
             const double velocityMetersPerSecond = Convert::knotsToMetersPerSecond(d->pluginSettings.getDefaultVelocity());
             const double seconds = distance / velocityMetersPerSecond;
             // Milliseconds
-            positionData.timestamp = previousPositionData.timestamp + std::round(seconds * 1000.0);
+            positionData.timestamp = previousPositionData.timestamp + static_cast<std::int64_t>(std::round(seconds * 1000.0));
         } else {
             positionData.timestamp = 0;
         }
@@ -261,7 +261,7 @@ void GpxParser::parseRoutePoint() noexcept
             const double velocityMetersPerSecond = Convert::knotsToMetersPerSecond(d->pluginSettings.getDefaultVelocity());
             const double seconds = distance / velocityMetersPerSecond;
             // Milliseconds
-            positionData.timestamp = previousPositionData.timestamp + std::round(seconds * 1000.0);
+            positionData.timestamp = previousPositionData.timestamp + static_cast<std::int64_t>(std::round(seconds * 1000.0));
         } else {
             positionData.timestamp = 0;
         }
@@ -342,7 +342,7 @@ inline void GpxParser::parseTrackPoint() noexcept
                 const double velocityMetersPerSecond = Convert::knotsToMetersPerSecond(d->pluginSettings.getDefaultVelocity());
                 const double seconds = distance / velocityMetersPerSecond;
                 // Milliseconds
-                positionData.timestamp = previousPositionData.timestamp + std::round(seconds * 1000.0);
+                positionData.timestamp = previousPositionData.timestamp + static_cast<std::int64_t>(std::round(seconds * 1000.0));
             } else {
                 positionData.timestamp = 0;
             }
